Remove $server parameter and move remaining parameter-logic up to client
Using two parameters with almost the same name is really confusing. With
the proper value-checking in place, we can reuse $servers for both
configuration modes.

Bonus: repair condition for if clause in client::config. I had it
switched around when removing the "case" blocks.

Bonus 2: Give localhost as the default server. Specifying an empty list
of servers as a default value was not a good idea. With the new default
value, the worst case will be that a server will wrongly try to connect
to itself, but the module will configure things coherently still.

Signed-off-by: Gabriel Filion <gabster@lelutin.ca>